@@ -1,74 +1,72 @@
 {
   "name": "webtorrent-desktop",
   "description": "WebTorrent, the streaming torrent client. For Mac, Windows, and Linux.",
-  "version": "0.17.2",
+  "version": "0.21.0",
   "author": {
     "name": "WebTorrent, LLC",
     "email": "feross@webtorrent.io",
     "url": "https://webtorrent.io"
   },
   "bugs": {
-    "url": "https://github.com/feross/webtorrent-desktop/issues"
+    "url": "https://github.com/webtorrent/webtorrent-desktop/issues"
   },
   "dependencies": {
-    "airplayer": "^2.0.0",
-    "application-config": "^1.0.0",
-    "arch": "^2.0.0",
-    "auto-launch": "^4.0.1",
-    "bitfield": "^1.0.2",
-    "capture-frame": "^1.0.0",
-    "chromecasts": "^1.8.0",
-    "cp-file": "^4.0.1",
-    "create-torrent": "^3.24.5",
-    "debounce": "^1.0.0",
-    "deep-equal": "^1.0.1",
+    "airplayer": "github:webtorrent/airplayer#fix-security",
+    "application-config": "^2.0.0",
+    "arch": "^2.1.1",
+    "auto-launch": "^5.0.5",
+    "bitfield": "^3.0.0",
+    "capture-frame": "^3.0.0",
+    "chokidar": "^3.2.2",
+    "chromecasts": "^1.9.1",
+    "create-torrent": "^4.4.1",
+    "debounce": "^1.2.0",
     "dlnacasts": "^0.1.0",
-    "drag-drop": "^2.12.1",
-    "es6-error": "^4.0.0",
+    "drag-drop": "^6.0.0",
+    "es6-error": "^4.1.1",
     "fn-getter": "^1.0.0",
-    "iso-639-1": "^1.2.1",
-    "languagedetect": "^1.1.1",
-    "location-history": "^1.0.0",
-    "material-ui": "^0.16.0",
+    "iso-639-1": "^2.1.0",
+    "languagedetect": "^2.0.0",
+    "location-history": "^1.1.1",
+    "material-ui": "^0.20.2",
     "mkdirp": "^0.5.1",
-    "musicmetadata": "^2.0.2",
-    "network-address": "^1.1.0",
-    "parse-torrent": "^5.7.3",
-    "prettier-bytes": "^1.0.1",
-    "react": "^15.2.1",
-    "react-dom": "^15.2.1",
-    "react-tap-event-plugin": "^1.0.0",
-    "rimraf": "^2.5.2",
-    "run-parallel": "^1.1.6",
-    "semver": "^5.1.0",
+    "music-metadata": "6.1.0",
+    "network-address": "^1.1.2",
+    "parse-torrent": "^7.0.1",
+    "prettier-bytes": "^1.0.4",
+    "prop-types": "^15.7.2",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "rimraf": "^3.0.0",
+    "run-parallel": "^1.1.9",
+    "semver": "^7.0.0",
     "simple-concat": "^1.0.0",
-    "simple-get": "^2.0.0",
-    "srt-to-vtt": "^1.1.1",
-    "vlc-command": "^1.0.1",
-    "webtorrent": "0.x",
-    "winreg": "^1.2.0",
-    "zero-fill": "^2.2.3"
+    "simple-get": "^3.1.0",
+    "srt-to-vtt": "^1.1.3",
+    "vlc-command": "^1.2.0",
+    "webtorrent": ">=0.107.16",
+    "winreg": "^1.2.4"
   },
   "devDependencies": {
-    "buble": "^0.14.0",
-    "cross-zip": "^2.0.1",
-    "depcheck": "^0.6.4",
-    "electron": "1.4.3",
-    "electron-osx-sign": "^0.3.0",
-    "electron-packager": "^8.0.0",
-    "electron-winstaller": "^2.3.0",
-    "gh-release": "^2.0.3",
+    "babel-eslint": "^10.0.3",
+    "buble": "^0.19.8",
+    "cross-zip": "^2.1.6",
+    "depcheck": "^0.9.0",
+    "electron": "~7.1.0",
+    "electron-osx-sign": "^0.4.14",
+    "electron-packager": "^14.0.6",
+    "electron-winstaller": "^4.0.0",
+    "gh-release": "^3.5.0",
     "minimist": "^1.2.0",
-    "nobin-debian-installer": "^0.0.10",
-    "nodemon": "^1.10.2",
-    "open": "0.0.5",
-    "plist": "^2.0.1",
-    "pngjs": "^3.0.0",
-    "run-series": "^1.1.4",
-    "spectron": "^3.3.0",
+    "nodemon": "^2.0.0",
+    "open": "^7.0.0",
+    "plist": "^3.0.1",
+    "pngjs": "^3.4.0",
+    "run-series": "^1.1.8",
+    "spectron": "^9.0.0",
     "standard": "*",
-    "tape": "^4.6.0",
-    "walk-sync": "^0.3.1"
+    "tape": "^4.11.0",
+    "walk-sync": "^2.0.2"
   },
   "engines": {
     "node": ">=4.0.0"
