@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const config = require('../src/config')
+const { CONFIG_PATH } = require('../src/config')
 const open = require('open')
 
-open(config.CONFIG_PATH)
+open(CONFIG_PATH)
