@@ -23,13 +23,13 @@ module.exports = {
   CRASH_REPORT_URL: 'https://webtorrent.io/desktop/crash-report',
   TELEMETRY_URL: 'https://webtorrent.io/desktop/telemetry',
 
-  APP_COPYRIGHT: 'Copyright © 2014-2016 ' + APP_TEAM,
+  APP_COPYRIGHT: 'Copyright © 2014-2019 ' + APP_TEAM,
   APP_FILE_ICON: path.join(__dirname, '..', 'static', 'WebTorrentFile'),
   APP_ICON: path.join(__dirname, '..', 'static', 'WebTorrent'),
   APP_NAME: APP_NAME,
   APP_TEAM: APP_TEAM,
   APP_VERSION: APP_VERSION,
-  APP_WINDOW_TITLE: APP_NAME + ' (BETA)',
+  APP_WINDOW_TITLE: APP_NAME,
 
   CONFIG_PATH: getConfigPath(),
 
@@ -70,11 +70,13 @@ module.exports = {
 
   DEFAULT_DOWNLOAD_PATH: getDefaultDownloadPath(),
 
-  GITHUB_URL: 'https://github.com/feross/webtorrent-desktop',
-  GITHUB_URL_ISSUES: 'https://github.com/feross/webtorrent-desktop/issues',
-  GITHUB_URL_RAW: 'https://raw.githubusercontent.com/feross/webtorrent-desktop/master',
+  GITHUB_URL: 'https://github.com/webtorrent/webtorrent-desktop',
+  GITHUB_URL_ISSUES: 'https://github.com/webtorrent/webtorrent-desktop/issues',
+  GITHUB_URL_RAW: 'https://raw.githubusercontent.com/webtorrent/webtorrent-desktop/master',
+  GITHUB_URL_RELEASES: 'https://github.com/webtorrent/webtorrent-desktop/releases',
 
   HOME_PAGE_URL: 'https://webtorrent.io',
+  TWITTER_PAGE_URL: 'https://twitter.com/WebTorrentApp',
 
   IS_PORTABLE: IS_PORTABLE,
   IS_PRODUCTION: IS_PRODUCTION,
