@@ -286,11 +286,16 @@ function getMenuTemplate () {
           click: () => windows.main.dispatch('pauseAllTorrents')
         },
         {
-          type: 'separator'
-        },
-        {
           label: 'Resume All',
           click: () => windows.main.dispatch('resumeAllTorrents')
+        },
+        {
+          label: 'Remove All From List',
+          click: () => windows.main.dispatch('confirmDeleteAllTorrents', false)
+        },
+        {
+          label: 'Remove All Data Files',
+          click: () => windows.main.dispatch('confirmDeleteAllTorrents', true)
         }
       ]
     },
@@ -303,6 +308,13 @@ function getMenuTemplate () {
           click: () => {
             const shell = require('./shell')
             shell.openExternal(config.HOME_PAGE_URL)
+          }
+        },
+        {
+          label: 'Release Notes',
+          click: () => {
+            const shell = require('./shell')
+            shell.openExternal(config.GITHUB_URL_RELEASES)
           }
         },
         {
@@ -321,13 +333,20 @@ function getMenuTemplate () {
             const shell = require('./shell')
             shell.openExternal(config.GITHUB_URL_ISSUES)
           }
+        },
+        {
+          label: 'Follow us on Twitter',
+          click: () => {
+            const shell = require('./shell')
+            shell.openExternal(config.TWITTER_PAGE_URL)
+          }
         }
       ]
     }
   ]
 
   if (process.platform === 'darwin') {
-    // Add WebTorrent app menu (Mac)
+    // WebTorrent menu (Mac)
     template.unshift({
       label: config.APP_NAME,
       submenu: [
@@ -346,8 +365,7 @@ function getMenuTemplate () {
           type: 'separator'
         },
         {
-          role: 'services',
-          submenu: []
+          role: 'services'
         },
         {
           type: 'separator'
@@ -370,8 +388,26 @@ function getMenuTemplate () {
       ]
     })
 
-    // Add Window menu (Mac)
-    template.splice(5, 0, {
+    // Edit menu (Mac)
+    template[2].submenu.push(
+      {
+        type: 'separator'
+      },
+      {
+        label: 'Speech',
+        submenu: [
+          {
+            role: 'startspeaking'
+          },
+          {
+            role: 'stopspeaking'
+          }
+        ]
+      }
+    )
+
+    // Window menu (Mac)
+    template.splice(6, 0, {
       role: 'window',
       submenu: [
         {
@@ -411,7 +447,7 @@ function getMenuTemplate () {
       })
 
     // Help menu (Windows, Linux)
-    template[4].submenu.push(
+    template[5].submenu.push(
       {
         type: 'separator'
       },
