@@ -33,15 +33,19 @@ function init (state, options) {
   const win = main.win = new electron.BrowserWindow({
     backgroundColor: '#282828',
     darkTheme: true, // Forces dark theme (GTK+3)
-    icon: getIconPath(), // Window icon (Windows, Linux)
-    minWidth: config.WINDOW_MIN_WIDTH,
-    minHeight: config.WINDOW_MIN_HEIGHT,
-    title: config.APP_WINDOW_TITLE,
-    titleBarStyle: 'hidden-inset', // Hide title bar (Mac)
-    useContentSize: true, // Specify web page size without OS chrome
-    show: false,
-    width: initialBounds.width,
     height: initialBounds.height,
+    icon: getIconPath(), // Window icon (Windows, Linux)
+    minHeight: config.WINDOW_MIN_HEIGHT,
+    minWidth: config.WINDOW_MIN_WIDTH,
+    show: false,
+    title: config.APP_WINDOW_TITLE,
+    titleBarStyle: 'hiddenInset', // Hide title bar (Mac)
+    useContentSize: true, // Specify web page size without OS chrome
+    width: initialBounds.width,
+    webPreferences: {
+      nodeIntegration: true,
+      enableBlinkFeatures: 'AudioVideoTracks'
+    },
     x: initialBounds.x,
     y: initialBounds.y
   })
@@ -56,7 +60,7 @@ function init (state, options) {
     win.setSheetOffset(config.UI_HEADER_HEIGHT)
   }
 
-  win.webContents.on('dom-ready', function () {
+  win.webContents.on('dom-ready', () => {
     menu.onToggleFullScreen(main.win.isFullScreen())
   })
 
@@ -72,27 +76,27 @@ function init (state, options) {
   win.on('hide', onWindowBlur)
   win.on('show', onWindowFocus)
 
-  win.on('enter-full-screen', function () {
+  win.on('enter-full-screen', () => {
     menu.onToggleFullScreen(true)
     send('fullscreenChanged', true)
     win.setMenuBarVisibility(false)
   })
 
-  win.on('leave-full-screen', function () {
+  win.on('leave-full-screen', () => {
     menu.onToggleFullScreen(false)
     send('fullscreenChanged', false)
     win.setMenuBarVisibility(true)
   })
 
-  win.on('move', debounce(function (e) {
+  win.on('move', debounce(e => {
     send('windowBoundsChanged', e.sender.getBounds())
   }, 1000))
 
-  win.on('resize', debounce(function (e) {
+  win.on('resize', debounce(e => {
     send('windowBoundsChanged', e.sender.getBounds())
   }, 1000))
 
-  win.on('close', function (e) {
+  win.on('close', e => {
     if (process.platform !== 'darwin') {
       const tray = require('../tray')
       if (!tray.hasTray()) {
@@ -137,37 +141,29 @@ function setAspectRatio (aspectRatio) {
 function setBounds (bounds, maximize) {
   // Do nothing in fullscreen
   if (!main.win || main.win.isFullScreen()) {
-    log('setBounds: not setting bounds because we\'re in full screen')
+    log('setBounds: not setting bounds because already in full screen mode')
     return
   }
 
   // Maximize or minimize, if the second argument is present
-  let willBeMaximized
-  if (maximize === true) {
-    if (!main.win.isMaximized()) {
-      log('setBounds: maximizing')
-      main.win.maximize()
-    }
-    willBeMaximized = true
-  } else if (maximize === false) {
-    if (main.win.isMaximized()) {
-      log('setBounds: unmaximizing')
-      main.win.unmaximize()
-    }
-    willBeMaximized = false
-  } else {
-    willBeMaximized = main.win.isMaximized()
+  if (maximize === true && !main.win.isMaximized()) {
+    log('setBounds: maximizing')
+    main.win.maximize()
+  } else if (maximize === false && main.win.isMaximized()) {
+    log('setBounds: unmaximizing')
+    main.win.unmaximize()
   }
 
+  const willBeMaximized = typeof maximize === 'boolean' ? maximize : main.win.isMaximized()
   // Assuming we're not maximized or maximizing, set the window size
   if (!willBeMaximized) {
-    log('setBounds: setting bounds to ' + JSON.stringify(bounds))
+    log(`setBounds: setting bounds to ${JSON.stringify(bounds)}`)
     if (bounds.x === null && bounds.y === null) {
       // X and Y not specified? By default, center on current screen
       const scr = electron.screen.getDisplayMatching(main.win.getBounds())
-      bounds.x = Math.round(scr.bounds.x + scr.bounds.width / 2 - bounds.width / 2)
-      bounds.y = Math.round(scr.bounds.y + scr.bounds.height / 2 - bounds.height / 2)
-      log('setBounds: centered to ' + JSON.stringify(bounds))
+      bounds.x = Math.round(scr.bounds.x + (scr.bounds.width / 2) - (bounds.width / 2))
+      bounds.y = Math.round(scr.bounds.y + (scr.bounds.height / 2) - (bounds.height / 2))
+      log(`setBounds: centered to ${JSON.stringify(bounds)}`)
     }
     // Resize the window's content area (so window border doesn't need to be taken
     // into account)
@@ -216,7 +212,7 @@ function toggleDevTools () {
   if (main.win.webContents.isDevToolsOpened()) {
     main.win.webContents.closeDevTools()
   } else {
-    main.win.webContents.openDevTools({ detach: true })
+    main.win.webContents.openDevTools({ mode: 'detach' })
   }
 }
 
