@@ -3,14 +3,14 @@ const ipcRenderer = require('electron').ipcRenderer
 
 const TorrentSummary = require('../lib/torrent-summary')
 const sound = require('../lib/sound')
-const {dispatch} = require('../lib/dispatcher')
+const { dispatch } = require('../lib/dispatcher')
 
 module.exports = class TorrentController {
   constructor (state) {
     this.state = state
   }
 
-  torrentInfoHash (torrentKey, infoHash) {
+  torrentParsed (torrentKey, infoHash, magnetURI) {
     let torrentSummary = this.getTorrentSummary(torrentKey)
     console.log('got infohash for %s torrent %s',
       torrentSummary ? 'existing' : 'new', torrentKey)
@@ -33,6 +33,7 @@ module.exports = class TorrentController {
     }
 
     torrentSummary.infoHash = infoHash
+    torrentSummary.magnetURI = magnetURI
     dispatch('update')
   }
 
@@ -62,7 +63,6 @@ module.exports = class TorrentController {
     torrentSummary.status = 'downloading'
     torrentSummary.name = torrentSummary.displayName || torrentInfo.name
     torrentSummary.path = torrentInfo.path
-    torrentSummary.magnetURI = torrentInfo.magnetURI
     // TODO: make torrentInfo immutable, save separately as torrentSummary.info
     // For now, check whether torrentSummary.files has already been set:
     const hasDetailedFileInfo = torrentSummary.files && torrentSummary.files[0].path
