@@ -3,7 +3,7 @@ const path = require('path')
 const prettyBytes = require('prettier-bytes')
 const React = require('react')
 
-const {dispatch, dispatcher} = require('../lib/dispatcher')
+const { dispatch, dispatcher } = require('../lib/dispatcher')
 
 const FlatButton = require('material-ui/FlatButton').default
 const RaisedButton = require('material-ui/RaisedButton').default
@@ -35,7 +35,7 @@ class CreateTorrentPage extends React.Component {
     // Then, exclude .DS_Store and other dotfiles
     const files = info.files
       .filter((f) => !containsDots(f.path, pathPrefix))
-      .map((f) => ({name: f.name, path: f.path, size: f.size}))
+      .map((f) => ({ name: f.name, path: f.path, size: f.size }))
     if (files.length === 0) return (<CreateTorrentErrorPage state={state} />)
 
     // Then, use the name of the base folder (or sole file, for a single file torrent)
@@ -65,9 +65,9 @@ class CreateTorrentPage extends React.Component {
     }
 
     // Create React event handlers only once
-    this.setIsPrivate = (_, isPrivate) => this.setState({isPrivate})
-    this.setComment = (_, comment) => this.setState({comment})
-    this.setTrackers = (_, trackers) => this.setState({trackers})
+    this.handleSetIsPrivate = (_, isPrivate) => this.setState({ isPrivate })
+    this.handleSetComment = (_, comment) => this.setState({ comment })
+    this.handleSetTrackers = (_, trackers) => this.setState({ trackers })
     this.handleSubmit = handleSubmit.bind(this)
   }
 
@@ -94,7 +94,8 @@ class CreateTorrentPage extends React.Component {
             marginBottom: 10
           }}
           hideLabel='Hide advanced settings...'
-          showLabel='Show advanced settings...'>
+          showLabel='Show advanced settings...'
+        >
           {this.renderAdvanced()}
         </ShowMore>
         <div className='float-right'>
@@ -104,12 +105,14 @@ class CreateTorrentPage extends React.Component {
             style={{
               marginRight: 10
             }}
-            onClick={dispatcher('cancel')} />
+            onClick={dispatcher('cancel')}
+          />
           <RaisedButton
             className='control create-torrent-button'
             label='Create Torrent'
             primary
-            onClick={this.handleSubmit} />
+            onClick={this.handleSubmit}
+          />
         </div>
       </div>
     )
@@ -141,9 +144,10 @@ class CreateTorrentPage extends React.Component {
           <label>Private:</label>
           <Checkbox
             className='torrent-is-private control'
-            style={{display: ''}}
+            style={{ display: '' }}
             checked={this.state.isPrivate}
-            onCheck={this.setIsPrivate} />
+            onCheck={this.handleSetIsPrivate}
+          />
         </div>
         <div key='trackers' className='torrent-attribute'>
           <label>Trackers:</label>
@@ -155,7 +159,8 @@ class CreateTorrentPage extends React.Component {
             rows={2}
             rowsMax={10}
             value={this.state.trackers}
-            onChange={this.setTrackers} />
+            onChange={this.handleSetTrackers}
+          />
         </div>
         <div key='comment' className='torrent-attribute'>
           <label>Comment:</label>
@@ -168,7 +173,8 @@ class CreateTorrentPage extends React.Component {
             rows={2}
             rowsMax={10}
             value={this.state.comment}
-            onChange={this.setComment} />
+            onChange={this.handleSetComment}
+          />
         </div>
         <div key='files' className='torrent-attribute'>
           <label>Files:</label>
@@ -191,9 +197,14 @@ function handleSubmit () {
     path: this.state.basePath,
     files: this.state.files,
     announce: announceList,
-    private: this.state.isPrivate,
     comment: this.state.comment.trim()
   }
+
+  // If torrent is not private, leave private flag unset. This ensures that
+  // the torrent info hash will match the result generated by other tools,
+  // including webtorrent-cli.
+  if (this.state.isPrivate) options.private = true
+
   dispatch('createTorrent', options)
 }
 
